gamestate: Rename activity node files.
