+++ libopenage/gamestate/activity/xor_event_gate.h
