+++ libopenage/gamestate/activity/xor_gate.h
