+++ libopenage/gamestate/entity_factory.cpp
@@ -15,10 +15,10 @@
 #include "event/event_loop.h"
 #include "gamestate/activity/activity.h"
 #include "gamestate/activity/end_node.h"
-#include "gamestate/activity/event_node.h"
 #include "gamestate/activity/start_node.h"
 #include "gamestate/activity/task_system_node.h"
-#include "gamestate/activity/xor_node.h"
+#include "gamestate/activity/xor_event_gate.h"
+#include "gamestate/activity/xor_gate.h"
 #include "gamestate/api/activity.h"
 #include "gamestate/component/api/idle.h"
 #include "gamestate/component/api/live.h"
