+++ libopenage/gamestate/system/activity.cpp
@@ -9,13 +9,13 @@
 #include "error/error.h"
 #include "log/message.h"
 
-#include "gamestate/activity/event_node.h"
 #include "gamestate/activity/node.h"
 #include "gamestate/activity/start_node.h"
 #include "gamestate/activity/task_node.h"
 #include "gamestate/activity/task_system_node.h"
 #include "gamestate/activity/types.h"
-#include "gamestate/activity/xor_node.h"
+#include "gamestate/activity/xor_event_gate.h"
+#include "gamestate/activity/xor_gate.h"
 #include "gamestate/component/internal/activity.h"
 #include "gamestate/component/types.h"
 #include "gamestate/game_entity.h"
